Prefer Memory<byte>
