+++ dr.ChromePasswordRecover/ICrypto.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace dr.ChromePasswordRecover
+{
+    public interface ICrypto
+    {
+        string DecryptString(Memory<byte> data);
+    }
+}
